Add auditing to get filter output
